Changed handling of on_project_created in tests. Instead of checking whether 'test' is in the command line args, we disconnect on_project_created from post_save in UnitTestBase.setUp and re-connect it in tests where we mock register_project_queues.
