+++ autograder/rest_api/apps.py
@@ -2,4 +2,7 @@ from django.apps import AppConfig
 
 
 class RestApiConfig(AppConfig):
-    name = 'rest_api'
+    name = 'autograder.rest_api'
+
+    def ready(self):
+        from . import signals
