+++ autograder/rest_api/views/ag_test_views/ag_test_command_views.py
@@ -69,14 +69,3 @@ class AGTestCommandDetailViewSet(TransactionRetrievePatchDestroyMixin, AGModelGe
     model_manager = ag_models.AGTestCommand.objects.select_related(
         'ag_test_case__ag_test_suite__project__course',
     )
-
-
-@receiver(post_save, sender=ag_models.AGTestCommand)
-def on_command_save(sender, instance: ag_models.AGTestCommand, created, **kwargs):
-    if not created:
-        clear_submission_results_cache(instance.ag_test_case.ag_test_suite.project_id)
-
-
-@receiver(post_delete, sender=ag_models.AGTestCommand)
-def on_command_delete(sender, instance: ag_models.AGTestCommand, *args, **kwargs):
-    clear_submission_results_cache(instance.ag_test_case.ag_test_suite.project_id)
